Consolidate some of the styling and components
