+++ packages/js/product-editor/src/components/add-products-modal/index.ts
@@ -1,2 +1,3 @@
 export * from './add-products-modal';
+export * from './reorder-products-modal';
 export * from './types';
