+++ packages/js/product-editor/src/components/add-products-modal/types.ts
@@ -1,10 +1,16 @@
 /**
  * External dependencies
  */
-import { Product } from '@woocommerce/data';
+import type { Product } from '@woocommerce/data';
 
 export type AddProductsModalProps = {
 	initialValue: Product[];
+	onSubmit( value: Product[] ): void;
+	onClose(): void;
+};
+
+export type ReorderProductsModalProps = {
+	products: Product[];
 	onSubmit( value: Product[] ): void;
 	onClose(): void;
 };
