--- packages/js/product-editor/src/components/reorder-products-modal/index.ts
@@ -1,2 +1,0 @@
-export * from './reorder-products-modal';
-export * from './types';
