--- packages/js/product-editor/src/components/reorder-products-modal/types.ts
@@ -1,10 +1,0 @@
-/**
- * External dependencies
- */
-import { Product } from '@woocommerce/data';
-
-export type AddProductsModalProps = {
-	products: Product[];
-	onSubmit( value: Product[] ): void;
-	onClose(): void;
-};
