+++ packages/js/product-editor/src/style.scss
@@ -39,7 +39,6 @@
 @import "components/attribute-control/attribute-skeleton.scss";
 @import "components/checkbox-control/style.scss";
 @import "components/add-products-modal/style.scss";
-@import "components/reorder-products-modal/style.scss";
 
 /* Field Blocks */
